Move GetOpts back to trees package
It is no longer used by storage.
